requirements.txt
selenium==4.9.0
pytest==7.2.2
allure-pytest==2.13.2
